requirements.txt
flask
opencv-python
numpy 
pillow 
requests 
beautifulsoup4
